+++ .dockerignore
@@ -1,0 +1,5 @@
+.env
+.DS_Store
+node_modules
+secondaryDB
+legacy_data
