+++ Dockerfile
@@ -1,0 +1,7 @@
+# syntax=docker/dockerfile:1
+FROM node:16.13.1
+WORKDIR /app
+COPY package.json /app/package.json
+RUN npm install
+COPY . /app
+CMD ["node", "server/index.js"]
