+++ package.json
@@ -21,10 +21,7 @@
   "homepage": "https://github.com/rpp36-sdc-utf8/rpp36-ziqian-service#readme",
   "dependencies": {
     "csv": "^6.2.0",
-    "csvtojson": "^2.0.10",
     "express": "^4.18.1",
-    "mongodb": "^4.8.1",
-    "mongoose": "^6.5.1",
     "morgan": "^1.10.0",
     "performance-now": "^2.1.0",
     "pg": "^8.7.3"
