+++ server/server.spec.js
@@ -1,9 +1,9 @@
 const request = require('supertest');
-const app = require('./server');
+const server = require('./server');
 
 describe('Test the root path', () => {
   test('It should response the GET method', () => {
-    request(app)
+    request(server)
       .get('/')
       .then((response) => {
         expect(response.statusCode).toBe(200);
